+++ examples/GW150914_flowMC.py
@@ -50,7 +50,7 @@ for ifo in ifos:
     data = Data.from_gwosc(ifo.name, start, end)
     ifo.set_data(data)
     psd_data = Data.from_gwosc(ifo.name, psd_start, psd_end)
-    ifo.set_psd(psd_data.to_psd(nperseg=data.duration * data.sampling_frequency))
+    ifo.set_psd(psd_data.to_psd(nperseg=int(data.duration * data.sampling_frequency)))
 
 # --- Waveform model ---
